Migrations: fix FK type mismatch in 032_article_editor_versions
UpdateService::runPendingMigrations aborted on production with
"SQLSTATE[HY000]: General error: 1005 Can't create table
sra_article_editor_versions (errno: 150 Foreign key constraint is
incorrectly formed)" — InnoDB rejected aev_article_fk and aev_user_fk
because the referencing columns were INT UNSIGNED while articles.id
and users.id are BIGINT UNSIGNED (001_users.sql line 9, 028_articles
.sql line 13).

Widened article_id and saved_by to BIGINT UNSIGNED. Local id stays
INT UNSIGNED — it's the table's own PK, not referenced from anywhere.

Edited the migration file in place. The runner only records migra-
tions on success (UpdateService.php:259), so the failed attempt
isn't logged in schema_migrations — re-running "Actualitzar" picks
up the corrected file and creates the table cleanly.

Added a comment block at the top of the SQL noting the type
contract so future migrations referencing articles or users don't
make the same mistake.

+++ database/migrations/032_article_editor_versions.sql
@@ -5,15 +5,22 @@
 -- dropdown of past versions and restore any of them back into TinyMCE.
 -- The current draft still lives on articles.editor_html (autosaved);
 -- this table only holds explicit snapshots the user wanted to keep.
+--
+-- FK type note: articles.id and users.id are both BIGINT UNSIGNED in
+-- this schema (see 001_users.sql and 028_articles.sql). The referencing
+-- columns here MUST match exactly or InnoDB rejects the constraint with
+-- errno 150 ("Foreign key constraint is incorrectly formed"). An
+-- earlier version of this migration used INT UNSIGNED and tripped that
+-- check on production installs.
 -- ─────────────────────────────────────────────────────────────────────────────
 
 CREATE TABLE `{prefix}article_editor_versions` (
-    `id`         INT UNSIGNED   NOT NULL AUTO_INCREMENT,
-    `article_id` INT UNSIGNED   NOT NULL,
-    `label`      VARCHAR(140)   NULL,
-    `html`       MEDIUMTEXT     NOT NULL,
-    `saved_by`   INT UNSIGNED   NULL,
-    `created_at` TIMESTAMP      NOT NULL DEFAULT CURRENT_TIMESTAMP,
+    `id`         INT UNSIGNED     NOT NULL AUTO_INCREMENT,
+    `article_id` BIGINT UNSIGNED  NOT NULL,
+    `label`      VARCHAR(140)     NULL,
+    `html`       MEDIUMTEXT       NOT NULL,
+    `saved_by`   BIGINT UNSIGNED  NULL,
+    `created_at` TIMESTAMP        NOT NULL DEFAULT CURRENT_TIMESTAMP,
     PRIMARY KEY (`id`),
     KEY `aev_article_idx` (`article_id`, `created_at`),
     CONSTRAINT `aev_article_fk` FOREIGN KEY (`article_id`)
